requirements.txt
# Core dependencies
flask==3.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-pinecone==0.0.3
pinecone-client==3.0.0
python-dotenv==1.0.0
requests==2.31.0

# Data processing
pypdf2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Additional utilities
click==8.1.7
gunicorn==21.2.0
